Build rockbox v3.15
Note that I had to use an older version of ubuntu with an older gcc
version because there were build errors in 22.04. The 14.04.1 was my
arbitrary pick (binutils 2.20.1 was released on 4 Mar 2010 and rockbox
3.15 — on 15 Nov 2019) and it worked!

It's awesome that docker hub still has the old image available for
download and the ubuntu repositories for this old version are still
available!